Schema DAO refactoring. Print LDD version in 'list-dd' command
