+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/data/LoadDataCmd.java
@@ -14,7 +14,7 @@ import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
 import gov.nasa.pds.registry.mgr.dao.DataLoader;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
-import gov.nasa.pds.registry.mgr.dao.SchemaUpdater;
+import gov.nasa.pds.registry.mgr.dao.schema.SchemaUpdater;
 import gov.nasa.pds.registry.mgr.util.ParamParserUtils;
