+++ src/test/java/tt/TestSchemaUpdater.java
@@ -3,7 +3,7 @@ package tt;
 import java.io.File;
 import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
-import gov.nasa.pds.registry.mgr.dao.SchemaUpdater;
+import gov.nasa.pds.registry.mgr.dao.schema.SchemaUpdater;
 import gov.nasa.pds.registry.mgr.dd.JsonLddLoader;
